Feat/ adding orderbook summary hash support
This PR aims to add the orderbook summary struct type and to provide a helper to calculate the hash of an orderbook.

To generate the hash of an orderbook it is needed a `OrderBookSummary` object.

This can be fetched from CLOB:

```python
orderbook = client.get_order_book("token id" )
```

Or built manually:

```python
orderbook = OrderBookSummary(
    market="0x...",
    asset_id="123...",
    bids=[],
    asks=[],
)
```

Then to generate the hash (if we pull the orderbook from clob the hash is included in the response) there are two options

```python
# using the client
hash = client.get_order_book_hash(orderbook)

# or using the util func
hash = generate_orderbook_summary_hash(orderbook)
```