Update get_balance_allowance.py
There is an error/typo in the no portion of this example. The Param was  not formatted properly. Specefically the token_id was not included inside the BalanceAlowanceParams. 
Resulting in the following error:
TypeError: get_balance_allowance() got an unexpected keyword argument 'token_id'


I resolved it by moving the token_id to the right place.